zakonczenie symulacji po zwolnieniu wszystkich fryzjerow

+++ kierownik.c
@@ -12,6 +12,9 @@
 #include <sys/prctl.h>
 #include <errno.h>
 
+// Licznik aktywnych fryzjerów
+volatile int aktywni_fryzjerzy = FRYZJERZY;
+
 void proces_kierownika() {
     srand(time(NULL));
     prctl(PR_SET_NAME, "kierownik", 0, 0, 0);
@@ -21,27 +24,42 @@ void proces_kierownika() {
         int akcja = rand() % 2; // 0 - zwolnienie fryzjera, 1 - ewakuacja klientów
 
         if (akcja == 0) {
-            int losowy_fryzjer = rand() % FRYZJERZY;
-            pid_t pid_fryzjera = getpid() + losowy_fryzjer + 1;
-            if(kill(pid_fryzjera, SIGTERM) == -1) {
-                perror("Błąd zwolnienia fryzjera");
-                fprintf(stderr, "Kod błędu errno: %d, opis błędu %s\n", errno, strerror(errno));
-                if (errno == ESRCH) {
-                    fprintf(stderr, "Proces o PID %d nie istnieje.\n", pid_fryzjera);
-                } else if (errno == EPERM) {
-                    fprintf(stderr, "Brak uprawnień do wysłania sygnału do procesu o PID %d.\n", pid_fryzjera);
-                }
+            if (aktywni_fryzjerzy > 0) {
+                int losowy_fryzjer = rand() % aktywni_fryzjerzy;
+                pid_t pid_fryzjera = getpid() + losowy_fryzjer + 1;
+                if (kill(pid_fryzjera, SIGTERM) == -1) {
+                    perror("Błąd zwolnienia fryzjera");
+                } else {
+                    printf("Kierownik: Zwolniono fryzjera o PID %d.\n", pid_fryzjera);
+                    aktywni_fryzjerzy--;
 
-            };
-            printf("Kierownik: Zwolniono fryzjera o PID %d.\n", pid_fryzjera);
+                    // Sprawdź, czy wszyscy fryzjerzy zostali zwolnieni
+                    if (aktywni_fryzjerzy == 0) {
+                        printf("Kierownik: Wszyscy fryzjerzy zostali zwolnieni. Kończę symulację.\n");
+                        kill(0, SIGINT); // Wysyła sygnał SIGINT do wszystkich procesów w grupie
+                        exit(0);
+                    }
+                }
+            }
         } else {
             printf("Kierownik: Ewakuacja klientów!\n");
+
+            // Zablokowanie wejścia do salonu
+            semctl(semafor, 4, SETVAL, 0);
+
+            ewakuacja = 1; // Ustawienie globalnej flagi ewakuacji
             for (int i = 0; i < KLIENCI; i++) {
                 pid_t pid_klienta = getpid() + FRYZJERZY + i + 1;
-                kill(pid_klienta, SIGUSR1); // Sygnał ewakuacji
+                kill(pid_klienta, SIGUSR1); // Wysłanie sygnału ewakuacji
             }
+
+            // Czekanie na zakończenie ewakuacji
+            sleep(5); // Czas na ewakuację klientów
+            printf("Kierownik: Ewakuacja zakończona.\n");
+
+            // Odblokowanie wejścia do salonu
+            semctl(semafor, 4, SETVAL, 1);
+            ewakuacja = 0; // Reset flagi ewakuacji
         }
     }
 }
-
-
